requirements.txt
streamlit
openai
python-dotenv
numpy
pandas
numpy-financial
fpdf
sqlalchemy